flask-login/requirements.txt
flask==1.0.2
python-dotenv~=0.10.1
flask-restplus==0.12.1
pytest-flask==0.10.0
sqlalchemy>=1.2.0b3
flask-sqlalchemy==2.3.2
Flask-Migrate==2.1.1
pymysql==0.9.3
pytest==3.6.2
py==1.5.3
pluggy==0.6.0
pytest-flask==0.10.0
pytest-cov==2.6.0
gunicorn==19.7.1
requests==2.18.4
Flask-Bcrypt==0.7.1
faker==1.0.4
sentry-sdk[flask]==0.7.7
flask-jwt-extended==3.20.0